+++ apps/probability/models/statistic/two_means_t_statistic.h
@@ -8,8 +8,8 @@ namespace Probability {
 class TwoMeansTStatistic : public TStatistic {
 public:
   void init(Data::SubApp subapp) override;
-  bool isValidParamAtIndex(int i, float p) override;
-  void setParamAtIndex(int index, float p) override;
+  bool isValidParamAtIndex(int i, double p) override;
+  void setParamAtIndex(int index, double p) override;
 
   void computeTest() override;
   void computeInterval() override;
@@ -25,29 +25,29 @@ protected:
   int numberOfStatisticParameters() const override { return k_numberOfParams; }
   enum ParamsOrder { X1, S1, N1, X2, S2, N2 };
   ParameterRepresentation paramRepresentationAtIndex(int i) const override;
-  float * paramArray() override { return m_params; }
+  double * paramArray() override { return m_params; }
   bool validateInputs() override;
 
 private:
-  float x1() { return m_params[ParamsOrder::X1]; }
-  float n1() { return m_params[ParamsOrder::N1]; }
-  float s1() { return m_params[ParamsOrder::S1]; }
-  float x2() { return m_params[ParamsOrder::X2]; }
-  float n2() { return m_params[ParamsOrder::N2]; }
-  float s2() { return m_params[ParamsOrder::S2]; }
+  double x1() { return m_params[ParamsOrder::X1]; }
+  double n1() { return m_params[ParamsOrder::N1]; }
+  double s1() { return m_params[ParamsOrder::S1]; }
+  double x2() { return m_params[ParamsOrder::X2]; }
+  double n2() { return m_params[ParamsOrder::N2]; }
+  double s2() { return m_params[ParamsOrder::S2]; }
 
-  float _xEstimate(float meanSample1, float meanSample2);
-  float computeT(float deltaMean,
-                 float meanSample1,
-                 float n1,
-                 float sigma1,
-                 float meanSample2,
-                 float n2,
-                 float sigma2);
-  virtual float computeDegreesOfFreedom(float s1, float n1, float s2, float n2);
-  virtual float computeStandardError(float s1, float n1, float s2, float n2);
+  double _xEstimate(double meanSample1, double meanSample2);
+  double computeT(double deltaMean,
+                 double meanSample1,
+                 double n1,
+                 double sigma1,
+                 double meanSample2,
+                 double n2,
+                 double sigma2);
+  virtual double computeDegreesOfFreedom(double s1, double n1, double s2, double n2);
+  virtual double computeStandardError(double s1, double n1, double s2, double n2);
 
-  float m_params[k_numberOfParams];
+  double m_params[k_numberOfParams];
 };
 
 }  // namespace Probability
